Moved flash cards to flash page, added small description on home page.
