+++ app/src/Pages/Home.tsx
@@ -7,29 +7,12 @@ The home page will contain the following components:
 1. A profile card that contains some info about Ivy
 2. A gallery of Ivy's previous work
  */
-const Animals = [
-    "Goat",
-    "Chicken",
-    "Donkey",
-    "Glubby",
-    "Rhino",
-    "Giraffe",
-    "LC",
-    "Ripley",
-    "Shrimp",
-    "Worm",
-    "Worms??"
-];
 const Home: React.FC = () => {
-    return (
-        <div className="justify-evenly sm:grid sm:gap-4 sm:grid-cols-2">
-        { Animals.map((animal) => (
-            <Card>
-                <p>{animal}</p>
-                <img className="rounded-2xl" alt="an awesome man" src="https://i.imgflip.com/50unao.png?a466320" />
-            </Card>
-        ))}
-        </div>
+    return(
+        <Card>
+            <h1>Hi! I'm Ivy</h1>
+            <p>I'm a tattoo artist working in Portland OR, and I specialize in creating creatures of all sorts for people of all sorts!</p>
+        </Card>
     )
 }
